+++ app/src/main/kotlin/at/bitfire/davdroid/sync/account/InvalidAccountException.kt
@@ -2,8 +2,11 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid
+package at.bitfire.davdroid.sync.account
 
 import android.accounts.Account
 
+/**
+ * Thrown when an account is invalid (usually because it doesn't exist anymore).
+ */
 class InvalidAccountException(account: Account): Exception("Invalid account: $account")
