+++ app/src/main/kotlin/at/bitfire/davdroid/ui/DebugInfoGenerator.kt
@@ -33,7 +33,7 @@ import androidx.work.WorkManager
 import androidx.work.WorkQuery
 import at.bitfire.dav4jvm.exception.DavException
 import at.bitfire.davdroid.BuildConfig
-import at.bitfire.davdroid.InvalidAccountException
+import at.bitfire.davdroid.sync.account.InvalidAccountException
 import at.bitfire.davdroid.R
 import at.bitfire.davdroid.TextTable
 import at.bitfire.davdroid.db.AppDatabase
